changed location of map
